ServerListPingEvent
An implementation of what could be done to make this event - not perfect in my opinion.
